stitch-v2: restructure cookbook into common/ + miles/ + slime/
Per feedback — the cookbook holds many things, layered: shared code factored out, the
framework/use-case specifics in per-framework subdirs, an experiment config per run.
Replaces the flat glm45_air_fp8 recipe (which over-collapsed and dropped slime).

- common/ (framework-agnostic, shared by every recipe): config.py (ModalConfig + mount
  paths), images.py (the weight-sync sglang SERVING image), server.py (register_server —
  the shared Server cls, parameterized by image/gpu/volumes/config), sidecar.py (the one
  serve() entrypoint), hooks.py (commit_and_wake/claim_pool/gated_rollout_request_hook
  logic), ray.py, process.py, smoke.py. Co-located hooks_test 4/4.
- miles/ (self-contained config, own pins): pins.py (trainer image + MILES_* pins),
  config.py (MilesConfig), app.py (EXPERIMENT_CONFIG -> common Server + miles Trainer +
  prep + entrypoints), prep.py (bf16/fp8/nvfp4 + torch_dist), hooks.py (re-export common),
  configs/glm45_air_fp8.py (migrated). torch-dist wrapper + megatron patch moved here.
- slime/ (symmetric): pins.py (slime image), config.py (SlimeConfig), app.py (download_model
  prep; publish hook key custom_delta_pre_push_path), hooks.py, configs/kimi_k2_6_int4.py.

Each framework's config is fully separate (own cli_args reflection). The Server is shared;
the Trainer + prep + pins are per-framework. The old *_disagg tree stays as reference
until e2e. Core suite 20/20; common hooks 4/4; miles/slime configs import (116/132 args).
